--- a/eslint.config.js
+++ b/eslint.config.js
@@ -18,6 +18,7 @@ export default tseslint.config([
       'tests/**/*',
       '**/*.test.{ts,tsx}',
       '**/*.spec.{ts,tsx}',
+      'server.js', // Express server with less strict typing
     ],
   },
   // Base configuration for all files
@@ -46,6 +47,11 @@ export default tseslint.config([
       '@typescript-eslint/explicit-module-boundary-types': 'off',
       '@typescript-eslint/no-explicit-any': 'warn',
       '@typescript-eslint/no-non-null-assertion': 'warn',
+      '@typescript-eslint/no-unsafe-assignment': 'warn',
+      '@typescript-eslint/no-unsafe-member-access': 'warn',
+      '@typescript-eslint/no-unsafe-call': 'warn',
+      '@typescript-eslint/no-unsafe-return': 'warn',
+      '@typescript-eslint/no-unsafe-argument': 'warn',
 
       // General rules
       'no-console': ['warn', { allow: ['warn', 'error'] }],