--- a/supabase/.temp/pooler-url
+++ b/supabase/.temp/pooler-url
@@ -1,0 +1,1 @@
+postgresql://postgres.hwlgbnhgoorlawloqpgh:[YOUR-PASSWORD]@aws-0-us-east-2.pooler.supabase.com:6543/postgres