… CLAUDE.md
… COMPREHENSIVE_PERFORMANCE_OPTIMIZATION_ROADMAP.md
… DATABASE_OPTIMIZATION_SUMMARY.md
… MANUAL_RLS_FIX_GUIDE.md
… MIGRATION_SUMMARY.md
… PERFORMANCE_OPTIMIZATIONS_COMPLETED.md
… PERFORMANCE_OPTIMIZATION_RESULTS.md
… RLS_PERFORMANCE_OPTIMIZATION.md
… SIMPLE_RLS_FIX.md
… STRIPE_PAYMENT_FIX.md
… docs/database-optimization.md
… docs/development-notes.md
… scripts/README.md
… scripts/apply-migration.md
… scripts/fix-rls-recursion.sql
… scripts/test-supabase-integration.ts
… server.js
… src/features/upload/FilePreviewGrid.tsx
… src/features/upload/FilePreviewItem.tsx
… src/lib/image-processor.ts
… supabase/.temp/gotrue-version
… supabase/.temp/postgres-version
… supabase/.temp/rest-version
… supabase/migrations/001_initial_schema.sql
… supabase/migrations/002_admin_usage_functions.sql
… supabase/migrations/003_add_admin_roles.sql
… supabase/migrations/003_add_admin_roles_fixed.sql
… supabase/migrations/004_fix_rls_recursion.sql
… supabase/migrations/005_temporary_fix.sql
… supabase/migrations/006_email_based_admin_functions.sql